+++ onert-micro/luci-interpreter/pal/mcu/PALFullyConnected.h
@@ -18,83 +18,36 @@
 #ifndef LUCI_INTERPRETER_PAL_FULLY_CONNECTED_H
 #define LUCI_INTERPRETER_PAL_FULLY_CONNECTED_H
 
-#include "Params.h"
-#include "PALUtils.h"
+#include "PALFullyConnectedCommon.h"
 
 namespace luci_interpreter_pal
 {
-template <typename InputType, typename WeightType, typename OutputType, typename BiasType>
-inline void FullyConnected(const FullyConnectedParams &params, const int32_t *input_shape,
-                           const InputType *input_data, const int32_t *filter_shape,
-                           const WeightType *filter_data, const BiasType *bias_data,
-                           const int32_t *output_shape, OutputType *output_data)
-{
-  const int32_t input_offset = params.input_offset;
-  const int32_t filter_offset = params.weights_offset;
-  const int32_t output_offset = params.output_offset;
-  const int32_t output_multiplier = params.output_multiplier;
-  const int output_shift = params.output_shift;
-  const int32_t output_activation_min = params.quantized_activation_min;
-  const int32_t output_activation_max = params.quantized_activation_max;
 
-  const int batches = input_shape[0];
-  const int output_depth = output_shape[1];
-  const int accum_depth = filter_shape[1];
-
-  for (int b = 0; b < batches; ++b)
-  {
-    for (int out_c = 0; out_c < output_depth; ++out_c)
-    {
-      BiasType acc = 0;
-      for (int d = 0; d < accum_depth; ++d)
-      {
-        int32_t input_val = input_data[b * accum_depth + d];
-        int32_t filter_val = filter_data[out_c * accum_depth + d];
-        acc += (filter_val + filter_offset) * (input_val + input_offset);
-      }
-      if (bias_data)
-      {
-        acc += bias_data[out_c];
-      }
-      int32_t acc_scaled = multiplyByQuantizedMultiplier(acc, output_multiplier, output_shift);
-      acc_scaled += output_offset;
-      acc_scaled = std::max(acc_scaled, output_activation_min);
-      acc_scaled = std::min(acc_scaled, output_activation_max);
-      output_data[out_c + output_depth * b] = static_cast<OutputType>(acc_scaled);
-    }
-  }
-}
 template <>
-inline void FullyConnected(const FullyConnectedParams &params, const int32_t *input_shape,
-                           const float *input_data, const int32_t *filter_shape,
-                           const float *filter_data, const float *bias_data,
-                           const int32_t *output_shape, float *output_data)
+inline void
+FullyConnected(const luci_interpreter_pal::FullyConnectedParams &params, const int32_t *input_shape,
+               const int8_t *input_data, const int32_t *filter_shape, const int8_t *filter_data,
+               const int32_t *bias_data, const int32_t *output_shape, int8_t *output_data)
 {
-  const float output_activation_min = params.float_activation_min;
-  const float output_activation_max = params.float_activation_max;
+  // MARK: At this moment this operation doesn't support
+  assert(false && "FullyConnected INT8 NYI");
+  (void)params;
+  (void)input_shape;
+  (void)input_data;
+  (void)filter_shape;
+  (void)filter_data;
+  (void)bias_data;
+  (void)output_shape;
+  (void)output_data;
+}
 
-  const int batches = input_shape[0];
-  const int output_depth = output_shape[1];
-  const int accum_depth = filter_shape[1];
-
-  for (int b = 0; b < batches; ++b)
-  {
-    for (int out_c = 0; out_c < output_depth; ++out_c)
-    {
-      float total = 0.f;
-      for (int d = 0; d < accum_depth; ++d)
-      {
-        total += input_data[b * accum_depth + d] * filter_data[out_c * accum_depth + d];
-      }
-      float bias_value = 0.0f;
-      if (bias_data)
-      {
-        bias_value = bias_data[out_c];
-      }
-      output_data[out_c + output_depth * b] =
-        std::min(std::max(total + bias_value, output_activation_min), output_activation_max);
-    }
-  }
+template <>
+inline void FullyConnected(const luci_interpreter_pal::FullyConnectedParams &, const int32_t *,
+                           const int16_t *, const int32_t *, const int8_t *, const int64_t *,
+                           const int32_t *, int16_t *)
+{
+  // MARK: At this moment this operation doesn't support
+  assert(false && "FullyConnected INT8 NYI");
 }
 
 } // namespace luci_interpreter_pal
