+++ onert-micro/luci-interpreter/pal/mcu/PALMul.h
@@ -18,96 +18,23 @@
 #ifndef LUCI_INTERPRETER_PAL_MUL_H
 #define LUCI_INTERPRETER_PAL_MUL_H
 
-#include "Params.h"
-#include "PALUtils.h"
-#include "ProcessBroadcastShapes.h"
+#include "PALMulCommon.h"
 
 namespace luci_interpreter_pal
 {
-template <typename T>
-inline void Mul(const ArithmeticParams &params, const int flat_size, const T *input1_data,
-                const T *input2_data, T *output_data)
-{
-  T activation_min, activation_max;
-  getActivationParams(params, &activation_min, &activation_max);
 
-  for (int i = 0; i < flat_size; ++i)
-    output_data[i] =
-      std::min(std::max(input1_data[i] * input2_data[i], activation_min), activation_max);
+template <>
+inline void Mul<int8_t>(const ArithmeticParams &, const int, const int8_t *, const int8_t *,
+                        int8_t *)
+{
+  assert(false && "Not IMPL yet");
 }
 
-template <typename T>
-inline void MulScalar(const ArithmeticParams &params, const int flat_size, const T *input_data,
-                      const T scalar_value, T *output_data)
+template <>
+inline void Mul<int16_t>(const ArithmeticParams &, const int, const int16_t *, const int16_t *,
+                         int16_t *)
 {
-  T activation_min, activation_max;
-  getActivationParams(params, &activation_min, &activation_max);
-
-  for (int i = 0; i < flat_size; ++i)
-    output_data[i] =
-      std::min(std::max(input_data[i] * scalar_value, activation_min), activation_max);
-}
-
-template <typename T>
-inline void
-BroadcastMul4DSlow(const ArithmeticParams &params,
-                   const luci_interpreter::RuntimeShape &input1_shape, const T *input1_data,
-                   const luci_interpreter::RuntimeShape &input2_shape, const T *input2_data,
-                   const luci_interpreter::RuntimeShape &output_shape, T *output_data)
-{
-  const int flat_size = input1_shape.flatSize();
-
-  if (params.broadcast_category == BroadcastableOpCategory::kScalarFirstBroadcast)
-  {
-    return MulScalar(params, flat_size, input2_data, input1_data[0], output_data);
-  }
-  else if (params.broadcast_category == BroadcastableOpCategory::kScalarSecondBroadcast)
-  {
-    return MulScalar(params, flat_size, input1_data, input2_data[0], output_data);
-  }
-
-  NdArrayDesc<4> desc1;
-  NdArrayDesc<4> desc2;
-  NdArrayDescsForElementwiseBroadcast(input1_shape, input2_shape, &desc1, &desc2);
-  const luci_interpreter::RuntimeShape extended_output_shape =
-    luci_interpreter::RuntimeShape::extendedShape(4, output_shape);
-
-  T activation_min, activation_max;
-  getActivationParams(params, &activation_min, &activation_max);
-
-  // In Tensorflow, the dimensions are canonically named (batch_number, row,
-  // col, channel), with extents (batches, height, width, depth), with the
-  // trailing dimension changing most rapidly (channels has the smallest stride,
-  // typically 1 element).
-  //
-  // In generated C code, we store arrays with the dimensions reversed. The
-  // first dimension has smallest stride.
-  //
-  // We name our variables by their Tensorflow convention, but generate C code
-  // nesting loops such that the innermost loop has the smallest stride for the
-  // best cache behavior.
-  for (int b = 0; b < extended_output_shape.dims(0); ++b)
-  {
-    for (int y = 0; y < extended_output_shape.dims(1); ++y)
-    {
-      for (int x = 0; x < extended_output_shape.dims(2); ++x)
-      {
-        for (int c = 0; c < extended_output_shape.dims(3); ++c)
-        {
-          const int output_data_offset =
-            ((b * extended_output_shape.dims(1) + y) * extended_output_shape.dims(2) + x) *
-              extended_output_shape.dims(3) +
-            c;
-
-          output_data[output_data_offset] =
-            std::min(std::max(input1_data[subscriptToIndex(desc1, b, y, x, c)] *
-                                input2_data[subscriptToIndex(desc2, b, y, x, c)],
-                              activation_min),
-                     activation_max);
-        }
-      }
-    }
-  }
+  assert(false && "Not IMPL yet");
 }
 
 } // namespace luci_interpreter_pal
